chatfinan/update_Data/requirements.txt
yfinance
pandas
apache-airflow==2.10.4
psycopg2-binary==2.9.10
pendulum